refactor: move OrgIdResolver to common/security
