+++ swatch-common-security/src/main/java/com/redhat/swatch/common/security/OrgIdResolver.java
@@ -18,12 +18,10 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package com.redhat.swatch.contract.resource;
+package com.redhat.swatch.common.security;
 
 import static com.redhat.swatch.common.security.RhIdentityHeaderAuthenticationMechanism.RH_IDENTITY_HEADER;
 
-import com.redhat.swatch.common.security.PskPrincipal;
-import com.redhat.swatch.common.security.RhIdentityPrincipalFactory;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.core.HttpHeaders;
