+++ swatch-contracts/src/main/java/com/redhat/swatch/contract/resource/api/v1/CapacityResourceV1.java
@@ -23,6 +23,7 @@ package com.redhat.swatch.contract.resource.api.v1;
 import com.redhat.swatch.common.model.HardwareMeasurementType;
 import com.redhat.swatch.common.model.ServiceLevel;
 import com.redhat.swatch.common.model.Usage;
+import com.redhat.swatch.common.security.OrgIdResolver;
 import com.redhat.swatch.configuration.registry.MetricId;
 import com.redhat.swatch.configuration.registry.ProductId;
 import com.redhat.swatch.configuration.registry.SubscriptionDefinitionGranularity;
@@ -42,7 +43,6 @@ import com.redhat.swatch.contract.repository.HypervisorReportCategory;
 import com.redhat.swatch.contract.repository.SubscriptionEntity;
 import com.redhat.swatch.contract.repository.SubscriptionRepository;
 import com.redhat.swatch.contract.resource.InMemoryPager;
-import com.redhat.swatch.contract.resource.OrgIdResolver;
 import com.redhat.swatch.contract.resource.ResourceUtils;
 import com.redhat.swatch.contract.resteasy.Page;
 import com.redhat.swatch.contract.resteasy.PageLinkCreator;
